Updated the MAEC Latest News page
Added 1 new news article about the upcoming working call on 2/1/17. Updated the intro text to remove the link to the blog, and refocused on it on MAEC Announce so the community to keep current on MAEC news in addition to this page. Also added news articles about the past working calls on 1/11/17 and 12/14/16.